chore: separate presets tests
